+++ lib/browser.ts
@@ -21,11 +21,13 @@ import * as resource_safety from './common/resource_safety';
 import * as io from './browser/io';
 import * as mqtt from './browser/mqtt';
 import * as crypto from './browser/crypto';
+import * as iot from './browser/aws_iot';
 
 export {
     io,
     mqtt,
     crypto,
+    iot,
     platform,
     resource_safety,
 };
