+++ lib/browser/io.ts
@@ -12,71 +12,8 @@
 * permissions and limitations under the License.
 */
 
-import { TlsVersion } from "../common/io";
-export * from "../common/io";
+export { TlsVersion } from "../common/io";
 
 export function is_alpn_available(): boolean {
     return false;
-}
-
-export class ClientBootstrap {
-    constructor() {
-    }
-}
-
-export class TlsContextOptions {
-    public min_tls_version: TlsVersion = TlsVersion.Default;
-    public ca_file?: string = undefined;
-    public ca_path?: string = undefined;
-    public alpn_list?: string = undefined;
-    public certificate_path?: string = undefined;
-    public private_key_path?: string = undefined;
-    public pkcs12_path?: string = undefined;
-    public pkcs12_password?: string = undefined;
-    public verify_peer: boolean = false;
-
-    override_default_trust_store(ca_path?: string, ca_file?: string): void {
-        this.ca_path = ca_path;
-        this.ca_file = ca_file;
-    }
-
-    static create_client_with_mtls(certificate_path: string, private_key_path: string): TlsContextOptions {
-        let opt = new TlsContextOptions();
-        opt.certificate_path = certificate_path;
-        opt.private_key_path = private_key_path;
-        opt.verify_peer = true;
-        return opt;
-    }
-
-    static create_client_with_mtls_pkcs(pkcs12_path: string, pkcs12_password: string): TlsContextOptions {
-        let opt = new TlsContextOptions();
-        opt.pkcs12_path = pkcs12_path;
-        opt.pkcs12_password = pkcs12_password;
-        opt.verify_peer = true;
-        return opt;
-    }
-
-    static create_server_with_mtls(certificate_path: string, private_key_path: string): TlsContextOptions {
-        let opt = new TlsContextOptions();
-        opt.certificate_path = certificate_path;
-        opt.private_key_path = private_key_path;
-        opt.verify_peer = false;
-        return opt;
-    }
-
-    static create_server_with_mtls_pkcs(pkcs12_path: string, pkcs12_password: string): TlsContextOptions {
-        let opt = new TlsContextOptions();
-        opt.pkcs12_path = pkcs12_path;
-        opt.pkcs12_password = pkcs12_password;
-        opt.verify_peer = false;
-        return opt;
-    }
-}
-
-export class ClientTlsContext {
-    readonly ctx_options: TlsContextOptions;
-
-    constructor(ctx_opt: TlsContextOptions) {
-        this.ctx_options = ctx_opt;
-    }
 }
